\name{Land plants}
\alias{landplants.fasta}
\alias{ex.lp.fasta}
\alias{ex.lp.fasta.rda}
\alias{short.landplants.phy}
\docType{data}
\title{Land plant rbcL sequences}
\description{
'landplants.fasta' is an example fasta file containing a 19,200 rbcL sequences of land plants (Embryophyta). 'ex.lp.fasta' and 'ex.lp.fasta.rda' is a subset of the landplants data with just 10 random plants in fasta format and as an R object(.rda).  'short.landplants' is another subset of the original data with 34 taxa, and is saved as a phylip formatted file.  
}
\format{fasta, R, phylip}
\keyword{datasets}
