\name{Data Sets}
\alias{DNA.fasta}
\alias{fasta_aa.aln}
\alias{fasta_dna.aln}
\alias{geno_test.tfam}
\alias{geno_test.tped}
\alias{phylip_interleaved_aa.aln}
\alias{phylip_interleaved_dna.aln}
\alias{PROTEIN.fasta}
\alias{simulation1.map}
\alias{simulation1.ped}
\docType{data}
\title{Data Sets}
\description{
The *.fasta data sets are sample data sets with only 10 sequences.  One is protein and one is DNA.  The .tfam/.tped files are transformed PLINK format.  The PLINK format is used mostly for Genome Wide Association Studies.
}
\format{fasta}
\keyword{datasets}

