\name{Internal functions}
\alias{find.input}
\alias{TestApp}
\alias{Error}
\alias{Renew}
\alias{RenewToken}
\alias{Time}
\alias{appINFO}
\alias{Check}
\alias{rplant.env}
\title{Internal Functions}
\description{
  All internal functions.
}
\usage{
TestApp(APP)
Error(ERR)
find.input(x)
RenewToken(print.curl=FALSE)
}
\arguments{
  \item{APP}{Name of application}
  \item{ERR}{Name of error}
  \item{x}{Name of input}
  \item{print.curl}{Prints the curl statement that can be used in the terminal, if curl is installed on your computer}
}

\keyword{internal}

