\name{Internal functions}
\alias{find.input}
\alias{TestApp}
\alias{Error}
\alias{Renew}
\title{Internal Functions}
\description{
  All internal functions.
}
\usage{
TestApp(APP)
Error(ERR)
Renew()
find.input(x)
}
\arguments{
  \item{APP}{Name of application}
  \item{ERR}{Name of error}
  \item{x}{Name of input}
}

\keyword{internal}

