\name{rplant-package}
\alias{rplant-package}
\alias{rplant}
\alias{rPlant}
\docType{package}
\title{rplant}
\description{
	R interface to iPlant's Discovery Enviornment and the Taxonomic Name Resolution Service (TNRS)
}
\details{
	\tabular{ll}{
	Package: \tab rplant\cr
	Type: \tab Package\cr
	Version: \tab 1.2\cr
	Date: \tab 2012-7-19\cr
	License: \tab GPL (>= 2)\cr}
}
\author{
Barb Banbury, Jeremy Beaulieu, Brian O'Meara
Maintainers: Barb Banbury <bbanbury@utk.edu>, Brian O'Meara <bomeara@utk.edu>
}
